% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavancier.R
\name{.lavancier_mult}
\alias{.lavancier_mult}
\title{.lavancier_mult}
\usage{
.lavancier_mult(v, s, B, C)
}
\arguments{
\item{v}{a real matrix.}

\item{s}{a positive number, the number of samples to take.}

\item{B}{logical indicating which model of vector should be returned.}

\item{C}{a logical vector of length \code{n} (the number of elements sampled and the number of columns of v), indicating for each draw if the probability weights for the next draw should be plotted.}
}
\value{
s samples from DSD(vv*)
}
\description{
.lavancier_mult
}
