% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavancier.R
\name{lavancier}
\alias{lavancier}
\title{lavancier}
\usage{
lavancier(v, s = 1, B = FALSE, C = rep(FALSE, 5000))
}
\arguments{
\item{v}{a matrix. See 'Details.'}

\item{s}{a positive number, the number of samples to take.}

\item{B}{logical indicating which model of vector should be returned. See 'Details.'}

\item{C}{a logical vector of length \code{n} (the number of elements sampled and the number of columns of v), indicating for each draw if the probability weights for the next draw should be plotted. See 'Details.'}
}
\value{
For \code{B = TRUE}, a vector of ones and zeros of length \code{N} (the number of elements to choose from), where ones indicate the elements drawn.

For \code{B = FALSE}, a vector of length \code{n} (the number of elements sampled) with elements drawn.
}
\description{
\code{lavancier} samples from fixed size determinantal sampling designs.
}
\details{
\code{v} has to be a matrix of size \code{Nxn} where \code{N} is the number of elements to choose from and \code{n} the number of elements sampled such that \code{vv*} is the kernel matrix of your sampling design. Two functions of this package provides such matrices. The function loonismary provides real matrices with a vector of first order inclusion probabilities as only argument. The function \code{periodicdsd} provides complex matrices with fixed size, equal first order probabilities, that exhibit some periodic behavior. Both models suit the function. See the examples.

If \code{s = 1}, the function returns a vector of length \code{n} of elements sampled.

If \code{s > 1}, the functions returns a dataframe of s columns in which the ith column contains the ith sample.

For \code{lavancier}, the default for \code{s} is \code{1}.
}
\examples{
lavancier(loonismary(c(0.5, 0.75, 0.75, 0.2, 0.4, 0.6, 0.8)))
#returns a vector of 4 elements sampled from 1:7.
lavancier(loonismary(c(0.5, 0.75, 0.75, 0.2, 0.4, 0.6, 0.8)), 5)
#returns a dataframe with 5 columns of 4 elements sampled from 1:7.
lavancier(loonismary(c(0.5, 0.75, 0.75, 0.2, 0.4, 0.6, 0.8)), B = TRUE)
#returns a vector of 4 elements sampled from 1:7
#and plots the evolution of the probability weights at every draw.

}
\references{
V., Loonis, X., Mary, (2017) "Determinantal Sampling Designs".
}
\seealso{
\link{loonismary} and \link{periodicdsd}.
}
