% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonisMary.R
\name{loonismary}
\alias{loonismary}
\title{loonismary}
\usage{
loonismary(Pi, K = F)
}
\arguments{
\item{Pi}{a vector of first order inclusion probabilities weights.}

\item{K}{a logical indicating which model of matrix should be returned. See 'Details.'}
}
\value{
For \code{K = TRUE}, the kernel matrix of the determinantal sampling design such that the first order inclusion probabilities are the ones given in \code{Pi}.

For \code{K = FALSE}, the matrix \code{V} such that \code{VV*} is the kernel matrix of the determinantal sampling design such that the first order inclusion probabilities are the ones given in \code{Pi}.
}
\description{
\code{loonismary} constructs a real projection matrix, kernels of fixed size sampling designs with first order inclusion probabilities given by \code{Pi}.
}
\details{
It is common in practice to work with fixed size sampling designs with prescribed first order inclusion probabilities. Constructing such a determinantal sampling design is equivalent to constructing a projection matrix with a prescribed diagonal. \code{loonismary} constructs this matrix.

As \code{Pi} is a vector of first order inclusion probabilities, it should be of length N, the number of elements \code{N} the set and should contain only values between 0 and 1 excluded such taht the sum of these values is \code{n}, the number of elements to be sampled.
}
\examples{
loonismary(c(0.5, 0.75, 0.75, 0.2, 0.4, 0.6, 0.8))
loonismary(rep(36/200, 200))

}
\references{
V., Loonis, X., Mary, (2017) "Determinantal Sampling Designs".
}
\seealso{
\link{lavancier}
}
