% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicdsd.R
\name{periodicdsd}
\alias{periodicdsd}
\title{periodicdsd}
\usage{
periodicdsd(n, r, N, K = FALSE)
}
\arguments{
\item{n}{a positive number, the number of items to be drawn, such that \code{n <= N}.}

\item{r}{a positive number, such that \code{r} and \code{N} are two relatively prime integers}

\item{N}{a positive number, the number of items to choose from.}

\item{K}{a logical indicating which model of matrix should be returned. See 'Details.'}
}
\value{
For \code{K = TRUE}, the kernel matrix of the determinantal sampling design.

For \code{K = FALSE}, the matrix \code{V} such that \code{VV*} is the kernel matrix of the determinantal sampling design.
}
\description{
\code{periodicdsd} constructs a complex matrix, kernel of a determinantal sampling designs such that first order inclusion probabilities are equal and the sampling design presents periodic behaviour. It is also possible to return a matrix \code{V} such that \code{VV*} is the kernel matrix.
}
\details{
\code{periodicdsd}, for periodic determinantal sampling design, constructs a toeplitz matrix constructed upon primitive Nth roots of the unity. This matrix then presents periodic behaviour.

Given that the first order inclusion probabilities are equal, these are equal \code{n/N}.
}
\examples{
periodicdsd(4, 2, 7, TRUE)
#returns a matrix K, kernel of the sampling design corresponding

periodicdsd(15, 1, 365)
#returns a matrix V

}
\references{
V., Loonis, X., Mary, (2017) "Determinantal Sampling Designs".
}
\seealso{
\link{lavancier}
}
